Diagnóstico: añadido viewPager a la página de resultados para mostrar posibles enfermedades
